chore: add js docs for context mutator hook
