+++ packages/react/src/index.ts
@@ -1,5 +1,6 @@
 export * from './evaluation';
 export * from './query';
 export * from './provider';
+export * from './context';
 // re-export the web-sdk so consumers can access that API from the react-sdk
 export * from '@openfeature/web-sdk';
