+++ packages/react/src/provider/index.ts
@@ -2,4 +2,3 @@ export * from './provider';
 export * from './use-open-feature-client';
 export * from './use-when-provider-ready';
 export * from './test-provider';
-export * from './use-context-mutator';
